fix: rename tag component #79
